organizing packages and error handling
